fix(doctor): isolate deferred module loaders
