+++ apps/www/components/ai/chat-new.tsx
@@ -15,11 +15,13 @@ import {
   useRouter,
 } from "@repo/internationalization/src/navigation";
 import { useMutation } from "convex/react";
+import { Effect } from "effect";
 import { useTranslations } from "next-intl";
 import { useTransition } from "react";
 import { AiChatModel } from "@/components/ai/chat-model";
 import { useAi } from "@/components/ai/context/use-ai";
 import { reportChatRuntimeError } from "@/components/ai/helpers/runtime-error";
+import { loadChatRuntime } from "@/components/ai/helpers/runtime-loader";
 import { useUser } from "@/lib/context/use-user";
 
 /** Renders the standalone new-chat input and starts the first message. */
@@ -65,7 +67,7 @@ export function ChatNew() {
           title: DEFAULT_TITLE,
           type: "study",
         }),
-        import("@/components/ai/helpers/runtime"),
+        Effect.runPromise(loadChatRuntime()),
       ]);
 
       const chatRuntime = createChatRuntime({
