+++ apps/www/components/shared/deferred-search-command.tsx
@@ -1,23 +1,10 @@
 "use client";
 
 import { useHotkeys } from "@mantine/hooks";
-import { Data, Effect } from "effect";
 import dynamic from "next/dynamic";
 
+import { loadSearchCommandModule } from "@/components/shared/search-command-module";
 import { useSearch } from "@/lib/context/use-search";
-
-/** Expected failure while warming the deferred command-search bundle. */
-class SearchCommandPreloadError extends Data.TaggedError(
-  "SearchCommandPreloadError"
-)<{
-  cause: unknown;
-  message: string;
-}> {}
-
-/** Loads command search only after a learner signals intent. */
-function loadSearchCommandModule() {
-  return import("@/components/shared/search-command");
-}
 
 const LazySearchCommand = dynamic(
   () => loadSearchCommandModule().then((module) => module.SearchCommand),
@@ -25,18 +12,6 @@ const LazySearchCommand = dynamic(
     loading: () => null,
     ssr: false,
   }
-);
-
-/** Warms command search without leaking a failed preload to the UI. */
-export const preloadSearchCommand = Effect.fn("www.search.preloadCommand")(() =>
-  Effect.tryPromise({
-    catch: (cause) =>
-      new SearchCommandPreloadError({
-        cause,
-        message: "Failed to preload command search.",
-      }),
-    try: loadSearchCommandModule,
-  }).pipe(Effect.ignore)
 );
 
 /** Registers lightweight shortcuts and mounts search after first activation. */
