+++ apps/www/components/sidebar/header-search.tsx
@@ -7,7 +7,7 @@ import { usePathname } from "@repo/internationalization/src/navigation";
 import { IconCommand, IconLetterK } from "@tabler/icons-react";
 import { Effect } from "effect";
 import { useTranslations } from "next-intl";
-import { preloadSearchCommand } from "@/components/shared/deferred-search-command";
+import { preloadSearchCommand } from "@/components/shared/search-command-module";
 import { useSearch } from "@/lib/context/use-search";
 
 /** Renders the header search trigger that opens the command search dialog. */
